arch/arm64: Fix return type of ffsl and flsl
The commit fixes the return type for ffsl and flsl which should return
an unsigned integer like ffs and fls. The promotion to unsigned long is
unnecessary (the builtin functions return int's anyway).

+++ arch/arm/arm64/include/uk/asm/atomic.h
@@ -68,7 +68,7 @@ static inline unsigned int ukarch_fls(unsigned int x)
  * @return The index of the least significant 1-bit of x, or if x is zero,
  *    the result is undefined. ffs(1)=0, ffs(3)=0, ffs(0x80000000)=31
  */
-static inline unsigned long ukarch_ffsl(unsigned long x)
+static inline unsigned int ukarch_ffsl(unsigned long x)
 {
 	return __builtin_ffsl(x) - 1;
 }
@@ -80,7 +80,7 @@ static inline unsigned long ukarch_ffsl(unsigned long x)
  * @return The index of the most significant 1-bit of x, or if x is zero,
  *   the result is undefined. fls(1)=0, fls(3)=1, fls(0x80000001)=31
  */
-static inline unsigned long ukarch_flsl(unsigned long x)
+static inline unsigned int ukarch_flsl(unsigned long x)
 {
 	return sizeof(x) * 8 - __builtin_clzl(x) - 1;
 }
